+++ grails2/plugin/release.sh
@@ -7,7 +7,7 @@ grails clean
 grails compile
 
 npm install
-grunt release
+gulp release-all
 
 #grails publish-plugin --snapshot --stacktrace
 grails publish-plugin --stacktrace